Change amount in getBytes to be max
The sensors return different length strings for different functions.
The lengths of those strings is not in the yaml definition. This commit
changes the exact number of bytes to read by calling getBytes to a
maximal amount. So either this maximal amount is read out or what ever
data is left. The function now only returns an error if no data is read.