Reduce buffer size limitation
Depending on the selected arduino board the i2c implementation uses
different buffer sizes. On small platforms some sensor commands cannot
be executed due to this restriction.
In order to avoid this limitation the responses are read out in several
steps using the maximal available buffer size.
The works fine for some sensors - but we cannot guarantee that it works
for all of them. In the carried out tests we have seen a sensor that
does not support read out of sensor data in several steps - but this
sensor has no issue with buffer sizes.